fix: balanced binary merkle tree merged proof (#6144)
Description
---
We now check if all the paths are correct. If they are not left
dangling, if they merge into one and that computes the root hash.

How Has This Been Tested?
---
`cargo test --package tari_mmr --lib --
balanced_binary_merkle_proof::test --nocapture`

What process can a PR reviewer use to test or verify this change?
---
Same test, but it would be nice to confirm my approach as well.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->