chore: use log4rs 1.3 (#6148)
Description
---
Switch to log4rs 1.3

Motivation and Context
---
https://github.com/tari-project/log4rs/pull/1
https://github.com/estk/log4rs/pull/345

It appears that the log4rs commits arent used since the cargo.lock rev
was set to commit
[`e1051fd3a1bec9c55d055f60176a96cf11e58505`](https://github.com/tari-project/log4rs/commit/e1051fd3a1bec9c55d055f60176a96cf11e58505)
https://github.com/tari-project/log4rs/commits/master/#:r3d:-listview-node-:r3e:
and the crate was renamed to tari_log4rs. This rename will only break
the current build if the cargo.lock is updated. Since no rev or tag is
used, current `master` would be used.

@SWvheerden Please verify the changes are in log4rs@1.3 that were
required. Just this right?
https://github.com/estk/log4rs/blob/c981ca4e2126c5728505c562b34926f9a9fad00b/Cargo.toml#L59

Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->