feat!: change proof of work to be dependant on target difficulty (#6156)
Description
---
Changes Tari to use Target difficulty rather than achieved difficulty
when looking at the chain with the most proof of work done.

Motivation and Context
---
Achieved difficulty can make it so that you get a lot of reorgs based on
blocks getting a lucky high achieved difficulty. This skews the results
in favour of a lucky block and can cause that lucky block to easily
reorg out many other blocks. It also allows an attacker to play games
with the block after they have found such a lucky block.

How Has This Been Tested?
---

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->