fix: wallet validation during reorgs (#6173)
Description
---
Fixes wallet validation during reorgs.

Motivation and Context
---
spent outputs that have been reorged, should not be marked as unspent.
They should be marked as unspen_mined_unconfirmed
If they are marked as unspent it means the wallet can immeadatly grab
the output and use it in a new transaction. The validation service first
needs to find out what the status of that output is, allthou the wallet
knows the original block the transaction was mined in was reorged out,
it needs to figure out if the transaction is still mined, unmined or
mempool.