feat: limit transaction size (#6154)
Description
---
Added checks to ensure that transactions negotiated between parties,
transactions to self and imported transactions do not exceed the RPC
frame size limit when the transaction is broadcast to the base node.
This can easily happen when lots of dust inputs are collected to be
spent. If a too-large transaction is detected in any of the transaction
service protocols it will be cancelled so that the user can try to
create a new transaction with different parameters, for example, to
reduce the recipient amount.

**Edit:**
- Added a wallet sqlite memory type connection to speed up tests with
large amounts of db activity
- Used fake outputs without valid bulletproof range proofs in the
`spend_dust` tests to speed it up even more as suggested by @SWvheerden

Motivation and Context
---
See #6108

How Has This Been Tested?
---
Added unit tests:
- `test_spend_dust_to_self_in_oversized_transaction`
- `test_spend_dust_to_other_in_oversized_transaction`
- `test_spend_dust_happy_path`

What process can a PR reviewer use to test or verify this change?
---
Code walk-through
Review unit tests

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->