feat: make the make_it_rain submission rate a float (#6180)
Description
---
- This PR will allow a wider transaction submission rate for the
make-it-rain tests.
- Also fixed cucumber tests where it needed to detect transactions with
a certain status. This was needed to fix the make-it-rain cucumber test.

Motivation and Context
---
A transaction submission fractional rate of between 1 and 2 txns/s is
required for extended stress tests.

How Has This Been Tested?
---
Cucumber test `Scenario: As a user I want to make-it-rain via command
line`

What process can a PR reviewer use to test or verify this change?
---
Cucumber test `Scenario: As a user I want to make-it-rain via command
line`

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->