refactor(hashing): move borsh hasher into tari_hashing crate (#6174)
Description
---
Move common hashers and domains required to interact with base layer
into common crate

Motivation and Context
---
In L2 we have a a borsh hasher that duplicates the functionality of the
consensus hasher in tari_core.
This crate allows us to use a common hasher implementation without
depending on tari_core.

TransactionHashDomain is also used in L1 and L2 so was moved into the
common crate.

How Has This Been Tested?
---
CI

What process can a PR reviewer use to test or verify this change?
---
CI

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->