Downgrade crossterm for Windows compatibility
A recent updates `crossterm` to the latest version, `0.27.0`, from version `0.25.0`. The
behaviour of keyboard events on Windows changed from version `0.26.0` upwards, making the
use of the console wallet and to a lesser degree the base node in Windows Console unusable
as multiple keystrokes events are generated for each keystroke.

This PR did not seek to find a solution for upgrading to the latest version of `crossterm`,
but merely to revert the change.