fix: downgrade crossterm for windows compatibility (#6204)
Description
---
A recent updates `crossterm` to the latest version, `0.27.0`, from
version `0.25.0`. The behaviour of keyboard events on Windows changed
from version `0.26.0` upwards, making the use of the console wallet and
to a lesser degree the base node in Windows Console unusable as multiple
keystrokes events are generated for each keystroke.

This PR did not seek to find a solution for upgrading to the latest
version of `crossterm`, but merely to revert the change.

Motivation and Context
---
See above.

How Has This Been Tested?
---
System-level testing

What process can a PR reviewer use to test or verify this change?
---
Review code changes
System-level testing

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

Co-authored-by: SW van Heerden <swvheerden@gmail.com>