feat: ledger key manager interface (#5644)
Description
---
This PR expands the console wallet and key manager interfaces to support
hardware wallets. It *does not* introduce functionality to communicate
with the hardware wallet, simply the interface to support it. Another PR
will be opened swapping out the placeholder software key manager for the
hardware key manager where appropriate.

Motivation and Context
---
So people can use their most excellent hardware wallets with the
MinoTari console wallet.

How Has This Been Tested?
---
Manually on osx

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

---------

Co-authored-by: C.Lee Taylor <leet@tari.com>