feat!: enable multiple second layers to be registered (#6225)
Description
---
Allows the registration of multiple second layers for validator nodes,
templates and burns. The default second layer (Tari) is specified by
leaving the `validator_network` field empty. For other networks, they
can either be private, where the private key is kept secret or public in
which case the `validator_network` private key can be exposed and
published.

Motivation and Context
---
Specifically for testing it will be useful to have multiple second layer
testnets run on the same base layer (I intend to use `igor`). For
example we may want to run a second layer "ContractNet" specifically for
users who wish to run templates and contracts but don't want to test
running validator nodes. In contrast, users who are interested in
testing validator nodes can run on the default second layer.

This way of doing it also allows users who wish to run their own private
networks, with their own validator nodes, to run them.

How Has This Been Tested?
---
Tested manually, and busy testing with changes to the second layer and
dan-testing

What process can a PR reviewer use to test or verify this change?
---
You can register templates and burns in the console wallet. Registering
validator nodes is a bit harder.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [ ] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [x] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->
BREAKING CHANGE: The merkle tree for validator nodes is calculated in a
new manner