ci: bump nightly version (#6241)
Description
---
Bumps the nightly toolchain to `nightly-2024-02-04` in preparation for a
curve library update.

Motivation and Context
---
An upcoming curve library update in #6228 will require moving the
nightly toolchain. In preparation, this PR moves it as far as it can go
without breaking. This lets us take care of lints now.

How Has This Been Tested?
---
Existing CI (but with the updated toolchain) passes.

What process can a PR reviewer use to test or verify this change?
---
Check that CI passes, that the nightly toolchain update is consistently
applied, and that all code changes reflect only lint fixes.