feat: allow wallet type from db to have preference (#6245)
Description
---
- Allow wallet type from the wallet database to have preference. This
will allow old wallets without wallet-type fields to be accepted as
software wallets when added to the config file.
- Set `newline_style = "Auto"` as we do not use JavaScript cucumber
anymore; this was an issue on Windows.

Motivation and Context
---
- Wallets prior to the wallet-type PR could not be used.
- Performing `cargo fmt` in Windows where the git preference has been
set to _check out as is and commit unix style_ forced `CRLF`. `Auto`
will give preference to the actual line-endings.

How Has This Been Tested?
---
System-level testing

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->