feat: improve node and wallet connection times (#6284)
Description
---
- Improved peer db stats to only update if the specific public address
was successfully connected to. This alleviates issues where the node
tries to connect to a peer's public address with a quality score > 0
where it should be 0.
- Update stats with ping-pong.
- Added a peer sync delay config option to let a preferred connection
get a head start in trying to connect.
- Discarded all peer identity claims without valid signatures
immediately when they are received.
- Reduced the minimum number of sync peers a wallet wants to connect to
from 50 to 8.

Motivation and Context
---
Peer db connection stats were not reflected correctly.
Wallet connection/re-connection times were slow.

How Has This Been Tested?
---
System-level testing

**Wallet re-connection times** with many pre-connected peers.


![image](https://github.com/tari-project/tari/assets/39146854/50effe9b-f554-4e48-8b84-72ea4152dd3b)


What process can a PR reviewer use to test or verify this change?
---
Code walk-through

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->