Add check to verify mempool state
Added a check to verify that the mempool last processed block is in sync with the
blockchain state whenever a new block template is requested, as adding a block to
the db and processing the same block in the mempool happens asynchronously. If this
is not done, miners can get block templates with double spends.