feat: add sent at timestamp to chat messages (#6314)
Description
---
Add sent at timestamp to messages. 

Motivation and Context
---
Because of the lack of a central time server, its hard to get accurate
timestamp, the chat services uses the received at as the timestamp.
But when the receiver gets messages through saf, it can get multiple at
the same time. This fudles up the order completely. So we add sent_at
timestamp to help with ordering of messages.