chore: refactor key manager hasher (#6330)
Description
---
Refactors the key manager hasher for clarity.

Closes #6327.

Motivation and Context
---
The key manager hasher is instantiated using a wrapper function that is
poorly named, as it is reused in multiple contexts. This PR refactors
for clarity.

How Has This Been Tested?
---
Tests pass.

What process can a PR reviewer use to test or verify this change?
---
Check the refactoring does not change behavior.