fix: make tx id random (#6380)
Description
---
Force tx_id to be random

Motivation and Context
---
If some apparent reason a transaction fails to be submitted after
creation using any method that does not supply a random tx_id, the tx_id
is created using the first output. But this will create
deterministically create a tx_id. This mean you can never rerun that
transaction as it will always create the same tx_id which results in a
db unique constraint failure.