feat: update ledger instructions (#6406)
Description
---
- Updated ledger instructions to build natively or via docker.
- Fixed compile errors.

Motivation and Context
---
See above.

How Has This Been Tested?
---
Performed a native build and install of the ledger app on Windows.

What process can a PR reviewer use to test or verify this change?
---
Physical build and install the ledger app.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->