chore: remove unnecessary commitment hashing (#6416)
Description
---
Removes unnecessary hashing of commitments used in faucet-related
signatures.

Motivation and Context
---
Signatures used for faucet functionality unnecessarily hash commitments
before passing them as an input message. This may have been done on the
assumption that the signing functions often refer to the message as a
challenge, despite this not being the case.

This PR removes the unnecessary hashing operations.

It may be worthwhile to refactor the associated signing functions to
avoid any possible confusion about the nature of messages and challenges
(as suggested in #6418).

How Has This Been Tested?
---
Existing tests pass.

What process can a PR reviewer use to test or verify this change?
---
Check that the commitment data is being passed into the signing
functions correctly, and that those functions do in face assume the
input is an arbitrary message and not a challenge.