chore: use safer slice splitting (#6437)
Description
---
Ensures that all slice splitting is properly checked to avoid panics.

Motivation and Context
---
There are a few remaining uses of
[`slice::split_at`](https://doc.rust-lang.org/std/primitive.slice.html#method.split_at)
that are used to split slices. Because this can panic on a bad index,
this PR moves each use to
[`slice::split_at_checked`](https://doc.rust-lang.org/std/primitive.slice.html#method.split_at_checked),
which cannot panic.

How Has This Been Tested?
---
Existing tests pass.

What process can a PR reviewer use to test or verify this change?
---
Ensure that each split is still being done at the same index, and that
the error handling is done properly.