feat: add private view key function (#6447)
Description
---
Adds a private view key function to transaction key manager service

Motivation and Context
---
We must be able to export the private view of the wallet. And to do this
we should not be required to expose the secret trait giving access to
all private keys in order to get this exported.