feat(wallet): add view key commands (#6426)
Description
---
Adds commands for exporting a view key and spend key for an existing
wallet, and also commands for creating a read-only wallet that only has
access to the view key.

Note this is only possible to the great work by @SWvheerden who enabled
this functionality in the wallet.

Motivation and Context
---
This allows you to keep an offline cold wallet that has spend access,
and at the same time listen for received funds and act on them. This
enables ecommerce and exchange support without the danger of running a
hot spending wallet.

How Has This Been Tested?
---
Tested locally

What process can a PR reviewer use to test or verify this change?
---
You can follow the steps in the [exchange
guide](https://github.com/tari-project/tari-dot-com/pull/231) to use all
of the additions in this PR

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->