feat: console wallet recovery improvements for ledger (#6455)
Description
---
When creating a ledger-based wallet don't show recovery seed words, as
they're not used to recover the ledger wallet type.
Additionally don't request recovery seed words for recovery, when
recovering a ledger based wallet.

Motivation and Context
---
It can be confusing as the seed words between the hardware device and
the software wallet won't actually align. In the event of hardware
wallet the software seed phrase can be ignored, and generated randomly
upon recovery. Only the ledger hardware recovery is important and that's
managed outside of the software wallet.

How Has This Been Tested?
---
Manually

What process can a PR reviewer use to test or verify this change?
---
Create a wallet
Recover a wallet


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify