feat: prepare MM backend stuff (#6448)
Description
---
This removes a few panics, and adds in a fake mainnet block.
Duplicates stagenet consensus values for mainnet. 

Motivation and Context
---
As we are getting closer to mainnet, we need to ensure that we can test
all of the features. This adds in a temporary mainnet block to use in
local testing. This block will not be the final one and will only be a
place holder.

How Has This Been Tested?
---
unit tests

---------

Co-authored-by: Hansie Odendaal <39146854+hansieodendaal@users.noreply.github.com>