feat: remove pre-mine create commands from the console wallet (#6451)
Description
---
Removed pre-mine create commands and associated code from the console
wallet.

Motivation and Context
---
This is a simple task and a console wallet is not needed to perform it.

How Has This Been Tested?
---
Code compiled successfully.
Unit tests pass.

What process can a PR reviewer use to test or verify this change?
---
Code review.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->